Move unit tests to unit subdirectory
