+++ frontend/mobile/__tests__/unit/selectors/authSelectors.test.ts
